Removed redundant code of the previous implementation of the list parsing.
