+++ src/django/api/constants.py
@@ -1,14 +1,3 @@
-class CsvHeaderField:
-    SECTOR = 'sector'
-    PRODUCT_TYPE = 'product_type'
-    SECTOR_PRODUCT_TYPE = 'sector_product_type'
-    COUNTRY = 'country'
-    NAME = 'name'
-    ADDRESS = 'address'
-    LAT = 'lat'
-    LNG = 'lng'
-
-
 class FileHeaderField:
     COUNTRY = 'country'
     NAME = 'name'
