+++ src/django/api/serializers/__init__.py
@@ -10,7 +10,6 @@ from .facility.facility_claim_details_serializer import FacilityClaimDetailsSeri
 from .facility.facility_claim_list_query_params_serializer import FacilityClaimListQueryParamsSerializer
 from .facility.facility_claim_review_note_serializer import FacilityClaimReviewNoteSerializer
 from .facility.facility_claim_serializer import FacilityClaimSerializer
-from .facility.facility_create_body_serializer import FacilityCreateBodySerializer
 from .facility.facility_create_query_params_serializer import FacilityCreateQueryParamsSerializer
 from .facility.facility_list_item_serializer import FacilityListItemSerializer
 from .facility.facility_list_items_query_params_serializer import FacilityListItemsQueryParamsSerializer
@@ -49,7 +48,6 @@ from .embed_config import EmbedConfigSerializer
 from .embed_fields import EmbedFieldsSerializer
 from .extended_field_list import ExtendedFieldListSerializer
 from .log_download_query_params import LogDownloadQueryParamsSerializer
-from .pipe_separated_field import PipeSeparatedField
 from .utils import (
     is_embed_mode_active,
     get_contributor_name,
