+++ src/django/api/views/__init__.py
@@ -44,7 +44,6 @@ from .facility.update_facility_activity_report_status import (
 )
 
 from .fields.create_embed_fields import create_embed_fields
-from .fields.create_nonstandard_fields import create_nonstandard_fields
 from .fields.nonstandard_fields_view_set import NonstandardFieldsViewSet
 
 from .tile.current_tile_cache_key import current_tile_cache_key
