+++ src/django/api/views/facility/facilities_view_set.py
@@ -78,7 +78,7 @@ from ...mail import send_claim_facility_confirmation_email
 
 from ...pagination import FacilitiesGeoJSONPagination
 from ...permissions import IsRegisteredAndConfirmed, IsSuperuser
-from ...sector_product_type_parser import SectorCache
+from ...sector_cache import SectorCache
 from ...serializers import (
     FacilityIndexSerializer,
     FacilityIndexDetailsSerializer,
