+++ src/django/contricleaner/lib/handlers/list_row_handler.py
@@ -15,7 +15,6 @@ class ListRowHandler(ABC):
 
     @abstractmethod
     def handle(self, rows: List[Dict]) -> ListDTO:
-        raise HandlerNotSetError("Next Handler wasn't set.")
         if self._next:
             return self._next.handle(rows)
