+++ src/react/src/util/propTypes.js
@@ -19,7 +19,6 @@ import {
     profileFieldsEnum,
     facilityListItemStatusChoicesEnum,
     ALLOW_LARGE_DOWNLOADS,
-    USE_OLD_UPLOAD_LIST_ENDPOINT,
     FEATURE,
     FEATURE_COLLECTION,
     POINT,
@@ -335,7 +334,6 @@ export const featureFlagPropType = oneOf([
     EMBEDDED_MAP_FLAG,
     EXTENDED_PROFILE_FLAG,
     ALLOW_LARGE_DOWNLOADS,
-    USE_OLD_UPLOAD_LIST_ENDPOINT,
 ]);
 
 export const facilityClaimsListPropType = arrayOf(
