+++ src/react/src/util/util.js
@@ -125,10 +125,7 @@ export const makeUserSignupURL = () => '/user-signup/';
 export const makeUserConfirmEmailURL = () =>
     '/rest-auth/registration/verify-email/';
 
-export const makeUploadFacilityListsURL = useOldUploadListEndpoint =>
-    useOldUploadListEndpoint
-        ? '/api/facility-lists/'
-        : '/api/facility-lists/createlistv2/'; // TODO: Remove this once testing of the parsing via ContriCleaner is complete.
+export const makeUploadFacilityListsURL = () => '/api/facility-lists/';
 export const makeFacilityListsURL = () => '/api/facility-lists/';
 export const makeSingleFacilityListURL = id => `/api/facility-lists/${id}/`;
 export const makeSingleFacilityListItemsURL = id =>
